Support case insensitive id in field
From comment https://github.com/forcedotcom/sf-fx-runtime-nodejs/pull/54#discussion_r672136053 it was decided that this function should accept an id in a case insensitive way.

This is implemented by creating a "map handler" that mutates all "id" key permutations to be lowercase. This handler also raises an error if more than one id key is provided.