Rework user-function tests and fixtures.
Node relies on this project's package.json module mode to determine what module
system `.js`  files use, when a `.js` file has no `package.json` of it's
own. Now that this project is in ESM module mode, it assumes all the
fixtures are also ESM. We need each CJS fixture to have a package.json
so we properly detect the fixture's module mode.