Update release docs (#186)
I was surprised to see this output while releasing:

```
npm notice total files:   24
npm notice
This operation requires a one-time password.
Enter OTP:
```

I wanted to document that a OTP will be requested and that it's from NPM. While contextually the output right above the prompt is from NPM, it also has a `npm notice`. 

It was initially unclear if this prompt was coming from NPM or from the next step in the release script.